+++ db/migrations/004_user_steam_accounts.rb
@@ -1,0 +1,13 @@
+Sequel.migration do
+  up do
+    alter_table :discord_users do
+      add_column :steam_account_id, Integer
+    end
+  end
+
+  down do
+    alter_table :discord_users do
+      drop_column :steam_account_id
+    end
+  end
+end