+++ CHANGELOG.md
@@ -8,6 +8,8 @@ glance - what to expact from upgrading to a new version.
 
 ### Added
 
+- Allow players to tell the bot about their Steam accounts.
+
 ### Changed
 
 ### Fixed